Alternative layout for config classes
